Transfers tests to new files, skips them
