+++ src/test/forms/output/BenefitsLines.test.js
@@ -2,36 +2,50 @@ import React from 'react';
 import { mount } from 'enzyme';
 import { cloneDeep, set } from 'lodash';
 
-import { StackedBarGraph } from '../../../forms/output/StackedBarGraph';
+import { BenefitsLines } from '../../../forms/output/BenefitsLines';
 import { CLIENT_DEFAULTS } from '../../../utils/CLIENT_DEFAULTS';
 
-jest.mock('react-chartjs-2', () => {
-  const BarMock = () => {
-    return null;
-  };
+// jest.mock('react-chartjs-2', () => {
+//   const LineMock = () => {
+//     return null;
+//   };
 
-  return {
-    Bar(props) {
-      return <BarMock { ...props } />;
-    },
-  };
-});
+//   return {
+//     Line(props) {
+//       return <LineMock { ...props } />;
+//     },
+//   };
+// });
 
-describe('<StackedBarGraph>', () => {
+describe.skip('<BenefitsLines>', () => {
+  let activePrograms;
   let client;
+  let defaultProps;
 
   const buildGraph = () => {
-    return mount(<StackedBarGraph client={ client } />);
+    return mount(<BenefitsLines { ...defaultProps } />);
   };
 
   beforeEach(() => {
+    activePrograms = [];
     client = cloneDeep(CLIENT_DEFAULTS);
+    defaultProps = {
+      activePrograms: activePrograms,
+      client:         client,
+      timescale:      'Monthly',
+      className:      'some-class',
+    };
+  });
+  
+  it('renders message when no benefits selected', () => {
+    const graph = buildGraph();
+
+    expect(graph.children).toHaveLength(1);
+    expect(graph.childAt(0).is('Message')).toBe(true);
   });
 
   it('renders with snap and current earned less than future earned', () => {
-    const benefits = [ 'snap' ];
-
-    set(client, 'current.benefits', benefits);
+    activePrograms.push('snap');
     set(client, 'current.earned', 100);
     set(client, 'future.earned', 200);
 
@@ -39,9 +53,7 @@ describe('<StackedBarGraph>', () => {
   });
 
   it('renders with snap and current earned greater than future earned', () => {
-    const benefits = [ 'snap' ];
-
-    set(client, 'current.benefits', benefits);
+    activePrograms.push('snap');
     set(client, 'current.earned', 200);
     set(client, 'future.earned', 100);
     
@@ -49,9 +61,7 @@ describe('<StackedBarGraph>', () => {
   });
 
   it('renders with section8', () => {
-    const benefits = [ 'section8' ];
-
-    set(client, 'current.benefits', benefits);
+    activePrograms.push('section8');
     set(client, 'current.earned', 100);
     set(client, 'future.earned', 200);
 
@@ -59,12 +69,8 @@ describe('<StackedBarGraph>', () => {
   });
 
   it('renders with both snap and section8', () => {
-    const benefits = [
-      'snap',
-      'section8',
-    ];
-
-    set(client, 'current.benefits', benefits);
+    activePrograms.push('snap');
+    activePrograms.push('section8');
     set(client, 'current.earned', 100);
     set(client, 'future.earned', 200);
